+++ src/map_data/point.rs
@@ -13,19 +13,19 @@ use super::rule::MapDataRule;
 #[derive(Clone, Serialize, Deserialize)]
 pub struct MapDataPoint {
     pub id: u64,
-    pub lat: f64,
-    pub lon: f64,
+    pub lat: f32,
+    pub lon: f32,
     pub lines: Vec<MapDataLineRef>,
     pub rules: Vec<MapDataRule>,
 }
 
 impl MapDataPoint {
-    pub fn distance_between(&self, point: &MapDataPointRef) -> f64 {
+    pub fn distance_between(&self, point: &MapDataPointRef) -> f32 {
         let self_geo = Point::new(self.lon, self.lat);
         let point_geo = Point::new(point.borrow().lon, point.borrow().lat);
         self_geo.haversine_distance(&point_geo)
     }
-    pub fn bearing_to(&self, point: &MapDataPointRef) -> f64 {
+    pub fn bearing_to(&self, point: &MapDataPointRef) -> f32 {
         let self_geo = Point::new(self.lon, self.lat);
         let point_geo = Point::new(point.borrow().lon, point.borrow().lat);
         self_geo.haversine_bearing(point_geo)
